+++ plugins/list-merging/CHANGELOG.md
@@ -2,3 +2,7 @@
 
 ## 0.2.1
 - update library to 0.1.1
+
+## 0.2.2
+- update library "anime-list" to 0.2.0
+- error messages are more readable